+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * `neuroconv.utils.jsonschema.get_schema_from_method_signature` can now support the `Dict[str, str]` typehint, which allows `DataInterface.__init__` and `.run_conversion` to handle dictionary arguments. [PR #360](https://github.com/catalystneuro/neuroconv/pull/360)
 * Added `neuroconv.tools.testing.data_interface_mixins` module, which contains test suites for different types of
   DataInterfaces [PR #357](https://github.com/catalystneuro/neuroconv/pull/357)
+* Added `expand_paths`. [PR #377](https://github.com/catalystneuro/neuroconv/pull/377)
 
 ### Testing
 * The tests for `automatic_dandi_upload` now follow up-to-date DANDI validation rules for file name conventions. [PR #310](https://github.com/catalystneuro/neuroconv/pull/310)
